moved AppRouteContentWrapper from demo to library
