+++ demo/src/index.tsx
@@ -20,15 +20,18 @@ const currentLanguage =
 const computedStyle = getComputedStyle(document.documentElement);
 const colorPrimary = computedStyle.getPropertyValue('--color-primary').trim();
 const colorSecondary = computedStyle.getPropertyValue('--color-secondary').trim();
+const colorContrastText = computedStyle.getPropertyValue('--color-button').trim();
 
 const theme = createTheme({
   ...locales[currentLanguage],
   palette: {
     primary: {
       main: colorPrimary,
+      contrastText: colorContrastText,
     },
     secondary: {
       main: colorSecondary,
+      contrastText: colorContrastText,
     },
   },
 });
