+++ library/src/components/index.ts
@@ -4,4 +4,5 @@ export { default as Create } from './Create';
 export { default as Edit } from './Edit';
 export { default as List } from './List/List';
 export { default as Detail } from './Detail';
+export * from './AppRouteContentWrapper/';
 export * from './layout';
